Reworking auth.service.ts and session.service.ts
Figuring out how to implement token based authentication. Not sure if we need to fetch the user and how we want to handle the persistent login etc. This is all a WIP and does not work yet. HTTP to API works.

TokenInterceptor.service.ts:
 Typo

auth.service.ts:
 changed methods from user based to token based authentication

session.service.ts:
 changed all methods to token based instead of userbased