Working on login and auth
This whole login thing is getting a pain in the xxx. For some reason Angular makes it very hard to get the API response and process the error. The API gives the correct answer when user creds are wrong but for some reason it cannot be found. ALso there is a new CORS error, this time with the /whoami endpoint.

login.component.ts: (WIP)
 rewrite login logic

auth.service.ts: (WIP)
 working on login

authGuard.service.ts:
 updates getSessionUser

session.service.ts:
 added sessionUser, we save that in the session serv now.