Working on login (client)
Finally got some way to get API error to frontend. Was fiddling a lot with this so there might be some parts which can be nuked lateron. For example the errurObj model.

Added a ErrorInterceptor which fetches all HTTP related errors. Added a MessageService to function as the bridge between the Interceptor and the component.
----
Changed:

errorObj.model.ts:
 created this model to hold errors but not sure if we want this.

ErrorInterceptor.service.ts:
 created this interceptor so we have one entry for all api errors

message.service.ts:
 added methods to it is usable. this is the bridge between interceptor and components.

auth.service.ts:
 removed all error related things, this script will only handle authentication stuff.

login.component.ts:
 changed.. a lot. Subscribed to message.service.ts so api errors do show now.

----
Next thing to do is figure out why API hates CORS (CORS errors) and where to save the session User so that it is accesable from all components (Auth?).