Trying more things to make this thing work
Still do not fully get the concept behind this http mechanism. Will continue to work this out. For now this is still WIP.

Trying in auth.service.ts something new i found on the web. Problem we have is that it is not possible (until now) to retrieve one value for the login process. The value is retrieved out of scope of the component.