HOTFIX 0.9.1.3
added a new component (probably temp) table that shows all registrations instead of only the recent ones. (hour-table-all-admin)
added new state to the state.service.ts so the admin can change to the new state.
changed time.service.ts to load both recent and all (THIS NEEDS TO BE CHANGED) for now a tmp fix
changed the selector of the hour-table-recent.component-admin.ts to app-hour-recent-table-admin.
changed the menu for the new button and state change to "OVERVIEW_ALL"

definitly need to optimize this soon tm.