Patch 0.9.1.4
Another small update. This update is a QOL improvement.

removed the page headers on admin.component.html (home) and home.component.html (user home) so we have more room. Also added some new padding to the content of the page.
removed the header from the menu.
changed timeservice so it will load data when the returned data.length === 0
changed nameplates on registration tables (admin)
added lastname to "user name" field on time table rows
added a icon in the manage-users.component.html table to make it easier to see if a user is active or not.