Update 0.9.3 Working on datatables
Fixed datatable issue where it goes bonkers when the data set changed. This was because the table was rendered before the http call was finished (which fetches the data).

Fixed this issue by rewriting some parts of the component and most of the time service. We do not want timeService to handle all of the fetching logic since we want to actually wait on the data to be fetched.

Because of this the next step will be rewriting fetching logic for the user components.

hour-table.component-all-admin.html:
 replaced *ngIf with [hidden] because the datatable does not play well with *ngIf. Big difference being that *ngIf actually does not render the HTML and hidden is just the css attribute that hides the HTML.

hour-table-all.component-admin.ts:
 restructured component logic so we can wait on the data to load. Also added some loading flags to implement a decent loading screen in the near future.

time.service.ts:
  simplifying the http methods, so they return a promise that we can use to finish the logic in the components that are using this service.