Update 0.9.3 Updating user components to use the new way of fetching time data.
Updating the user components to use the new way we fetch data (see prev commit). We are moving logic to the component itself so we can do stuff when the data is actually fetched. This makes the logic less complicated and bloated since we only need that specific data for a specific component for a specific user role.

Also added some todo's.

hour-form.component.ts:
 changed submitTime() to use the new way of fetching data.

time.service.ts:
 some small changes. We need to keep the recentTimesChanged mechanism here since the 'recentTimes' table for the user is a seperate component (seperate from hour-table) that depends on this mechanism.