Update 0.9.3.1 Added options to datatable
Added an option to the datatables options so that the table will sort on date on render. Also added the option 'saveState: true' so that the filter is saved on rerender, resulting in the same dataset before rerender. Added ANOTHER option 'autoWidth: false' to prevent the datatable from resizing the columns on rerender.