Update 0.9.5.0: Working on Excel export function
Added logic to extract a usable dataset from the datatable component (aka dt). The dtInstance contains a lot of application data and the registration data fetched from the server.

We now extract the currently displayed dataset (including all pages) with search applied. This gives the user the possibility to filter the dataset.

Next step is to use the exceljs library to put the data into a readable excel sheet and present the file to the user as a download.

We need to make changes (according to the internet) to the tsConfig. Exceljs is by default 'compiled' for the server side and we want to use it client side. Therefore we need to add a line to the tsConfig that tells Angular to compile this library with the frontend in mind.