Update 0.9.5.0: Working on Excel export function 2
Done most of the needed logic to export currently selected data to xlsx format.

Added a new service that does the data to xlsx bit. Need to make this more general and make sure xlsx configs are passed from the place where the data comes from (in this case from the hour-table) so we can reuse this service everywhere we need a export function.