Update v0.9.6.0: Working on mail functionality
Added a mailer service that is responsible for sending e-mails when triggered.

Added checkboxes to the manage-users.component.html which are visible when creating users or changing a password of a user. These checkboxes will trigger the mailer service to try and send an e-mail.

Added logic to the manage-users.component.ts to trigger the mailerservice.