Update v0.9.6.0: QOL improvements and general changes
Added a new component and installed it. The Dashboard component. This still needs all the work ofc. For now a placeholder.
Updated state service to support the new dashboard state.
Changed icons of the menu to more sensible ones. Since we are adding a dashboard we needed the chart icon for that.
Added checkmark icons on the columns 'Akkoord' so it looks more sensible when something is not checked yet. Done this for all time tables.