Update v0.9.8: Issue #8 Fixed the double call and error response.
Issue was the onSubmit method in the login component. It called the whoami in the callback on the login().then() (promise) without checking if login succeeded.

Added a method to get the token value which will only be available when login succeeds since the server needs to generate the token (a JWT token) and will only return a token when authentication is succesfull.