Update v0.9.8: Ticket #16 Working on the Setting service and initializing it when the app loads.
The Setting service gets initialised when the app starts. Data get's fetched when the application is done loading.

~ app.module.ts:
added the settingservice to the providers array. added a provider with the APP_INITIALIZER type to make this run in init. This thing also needs a method to call when the app gets initialized so that is why "useFactory" is therefor. The name is somewhat funky since we are not calling a factory at this point.

+ setting.model.ts:
default model for the setting thing.

+ setting.service.ts:
service responsible for doing the api call to the server and to hold the setting data.