Update v0.9.8: Ticket #16 After rewriting the setting service many times i finally got it to work and implemented it properly.
The thing only gets loaded once (singleton service) and the data gets loaded in before the route is routed. But after authentication because auth is good.

Menu looked like a good place to start implementng this.

~ app.component.ts:
added settingService globally.

~ app.module.ts:
same

~ app.routing.module.ts:
added the settingDataResolver to the admin route. This also needs to be added to the home route. Makes the service load before loading the route.

~ menu.component.html:
added setting checks on buttons.

~ menu.component.ts:
after hours of figuring out how to implement the setting shite. I went for this approach on the menu thing. Guess we can use this method in other places aswel.

~ setting.service.ts:
simplified. dont want to talk about this. Previous version was glorious.

~ state.service.ts:
todo

+ SettingDataRevoler:
the first data resolver. We learned new angular things today.