Update v0.9.8: Ticket #16 Alright saving all the settings at once was a PITA to make it work with springboot.
It now works so we can continue to implement the settings in the application to alter its logic and shite.

~ menu.component.html:
added permission checks on the buttons and fuqing around with highlighting the button that is currently active.

~ settings.component.ts:
Made the onSubmit actually submit the changes and not just the data we retrieved from the serv.

~ state.service.ts:
Did some rewriting making use on static constants that we are now referencing everywhere instead of the magic strings everywhere. This standarizes things somewhat.