Update v0.9.8: [ENVIRONMENTS] Adding new environments so we can do ng build configuration=env. Did refactoring of img file names because i did not want any capitals in that.
Added Bimmate logo and configuration files.

~ angular.json:
added configurations for luca and bimmate. These configs are used when building the angular application with the ng build command. The param passed to '--configuration' points to these config blocks in this file. These config blocks tells the builder many things but it also tells the builder to use a specific environment file.

~ environment.*.ts:
added new file for bimmate. added logo url to the env file so we can load a specific logo for each env.