Patch v0.9.8.2: Fix where the breaktimes are not loaded when admin does not have the permission to register worktimes (setting ADMIN_REGISTER_TIME).
hour-table-all.component-admin.ts was missing the timeservice init method and because we never tested this where the ADMIN_REGISTER_TIME setting was false, the register time component got loaded first. Which explains why the breaktimes got loaded anyway because the timeservice is a singleton thus only one instance ia available for all components.