UPDATE v0.9.9 #35: Partial rewrite of the charData component to be able to switch out the data when new data is fetched.
~ chartdata.service.ts:
Moved the subscribe to the contructor since these actually need to be declared once. This is actually just a callback function and will get called for every emit.

Updated the fetchChartData so it uses the new endpoint which is just for fetching chartData.

Added custom method to get colors for the bars and lines in the charts. This was needed because we are now creating the charts on component level and they are not "declared" in the HTML directly.

~ dashboard.component.ts:
Added a method to create the charts. We are now creating the charts on the component level instead of the template and leaving it to the HTML to do all the work. This means we need to introduce some variables that can hold the charts and such. Initially this was done so we can trigger the charts to do a rerender but that is not needed. Still this seems somewhat better then leaving this reponsibility to the HTML template.
Moved the subscribe on the chartData emitter to the contructor because we need to reuse this callback during data fetching.

Added logic to the onsubmit and reset buttons of the filter element.

Reworked the getExpectedGrossIncome() method since that was a mess.