UPDATE v0.9.9 #35: Bug fixes and adding support for one filter selection (ex: just location).
~ chartdata.service.ts:
added logic in the fetchChartData() method to handle requests with one of the parameters. Where one param is null.

~ dashboard.component.html:
fixed a derp where the method getExpectedGrossIncome() got called in the template. It should just bind the result in statsExpectedGrossIncome, this saves one method call.

~ dashboard.component.ts:
fixed charts being rendered multiple times by adding a destory call when they are rendered.

fixed the onSubmit() by adding a submitDisabled flag. Which shows a toast when true instead of calling the api again. Thisway we ensure we have all location and year data before letting the user filter and call the api. Otherwise we do not have all the years due to the way things work. It requires a pretty extensive rework of the component to work around that. So this is the quickest fix.

fixed expectedgross income calculation being incorrect after filtering on location. Added a check where we always get the rate of the filterLocationSelection is that is not set to null.