UPDATE v0.9.9 #29: Adding checkmark to show/hide inactive users.
~ manage-users.component.html:
added new input checkmark on the view which is bound to a component level variable.

~ manage-users.component.css:
added some custom styling for the checkmark input field on the view.

~ manage-users.component.ts:
updated the logic for retrieving users and the checkmark logic which get called in the onChange event of the checkmark input field.

~ user.service.ts:
updated data fetching. Removed unused emitter. Added new method for refreshing the data from cache when the checkmark input field changes.